+++ vitest.config.ts
@@ -20,6 +20,7 @@ export default defineConfig({
     },
     resolve: {
         alias: {
+            '@sigx/i18n/server/node': resolve(__dirname, 'packages/i18n/src/server-node.ts'),
             '@sigx/i18n/server': resolve(__dirname, 'packages/i18n/src/server.ts'),
             '@sigx/i18n/vite': resolve(__dirname, 'packages/i18n/src/vite.ts'),
             '@sigx/i18n': resolve(__dirname, 'packages/i18n/src/index.ts')